Fix RecordingProvider package + BC test import after the move
The preceding commit moved the file but, due to a staging mistake, kept
the old 'package ...jdk17plus' declaration and omitted the content edits,
so it did not compile on its own. Apply the package change to
com.amazon.corretto.crypto.provider.test, the JDK 8-safe
Provider(String, double, String) constructor and explicit Map.Entry type
(no 'var'), and the explicit RecordingProvider import in the BouncyCastle
JSSE ML-KEM test.
